+++ src/app/features/tipo-documento/tipo-documento.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { EntityTableComponent } from '../../shared/components/entity-table/entity-table.component';
 import { FormModalComponent } from '../../shared/components/form-modal/form-modal.component';
 import { NgIf } from '@angular/common';
-import {TipoDocumento} from '../../core/models/tipoDocumento.model';
-import {TipoDocumentoService} from '../../core/services/tipo-documento.service';
+import { TipoDocumento } from '../../core/models/tipoDocumento.model';
+import { TipoDocumentoService } from '../../core/services/tipo-documento.service';
+import { NotificationService } from '../../shared/service/notification.service';
 
 @Component({
   selector: 'app-tipo-documento',
@@ -21,7 +22,6 @@ export class TipoDocumentoComponent implements OnInit {
   isModalOpen = false;
   tipoDocumentoSeleccionado: TipoDocumento | null = null;
   columns = [
-    { key: 'id', label: 'ID' },
     { key: 'codigo', label: 'Código' },
     { key: 'nombre', label: 'Nombre' },
     { key: 'descripcion', label: 'Descripción' },
@@ -34,7 +34,10 @@ export class TipoDocumentoComponent implements OnInit {
     { key: 'estado', label: 'Estado', type: 'checkbox' }
   ];
 
-  constructor(private tipoDocumentoService: TipoDocumentoService) {}
+  constructor(
+    private tipoDocumentoService: TipoDocumentoService,
+    private notificationService: NotificationService
+  ) {}
 
   ngOnInit(): void {
     this.loadTiposDocumento();
@@ -42,8 +45,6 @@ export class TipoDocumentoComponent implements OnInit {
 
   loadTiposDocumento(): void {
     this.tipoDocumentoService.getAll().subscribe(data => {
-      console.log('Datos obtenidos del backend:', data);
-
       this.tiposDocumento = data;
     });
   }
@@ -61,23 +62,41 @@ export class TipoDocumentoComponent implements OnInit {
     if (this.tipoDocumentoSeleccionado) {
       this.tipoDocumentoService.update(this.tipoDocumentoSeleccionado.id, tipoDocumento).subscribe(() => {
         this.loadTiposDocumento();
+        this.notificationService.showSuccess('Tipo de Documento actualizado con éxito');
+      }, error => {
+        console.error('Error al actualizar el Tipo de Documento:', error);
+        this.notificationService.showError('Error al actualizar el Tipo de Documento');
       });
     } else {
       this.tipoDocumentoService.create(tipoDocumento).subscribe(() => {
         this.loadTiposDocumento();
+        this.notificationService.showSuccess('Tipo de Documento creado con éxito');
+      }, error => {
+        console.error('Error al crear el Tipo de Documento:', error);
+        this.notificationService.showError('Error al crear el Tipo de Documento');
       });
     }
     this.cerrarModal();
   }
 
-  editarTipoDocumento(tipoDocumento: TipoDocumento): void {
-    this.tipoDocumentoSeleccionado = tipoDocumento;
-    this.abrirModal();
+  async editarTipoDocumento(tipoDocumento: TipoDocumento): Promise<void> {
+    const confirmed = await this.notificationService.showConfirmation('¿Desea editar este Tipo de Documento?');
+    if (confirmed) {
+      this.tipoDocumentoSeleccionado = tipoDocumento;
+      this.abrirModal();
+    }
   }
 
-  eliminarTipoDocumento(tipoDocumento: TipoDocumento): void {
-    this.tipoDocumentoService.delete(tipoDocumento.id).subscribe(() => {
-      this.loadTiposDocumento();
-    });
+  async eliminarTipoDocumento(tipoDocumento: TipoDocumento): Promise<void> {
+    const confirmed = await this.notificationService.showConfirmation('¿Desea eliminar este Tipo de Documento?');
+    if (confirmed) {
+      this.tipoDocumentoService.delete(tipoDocumento.id).subscribe(() => {
+        this.loadTiposDocumento();
+        this.notificationService.showSuccess('Tipo de Documento eliminado con éxito');
+      }, error => {
+        console.error('Error al eliminar el Tipo de Documento:', error);
+        this.notificationService.showError('Error al eliminar el Tipo de Documento');
+      });
+    }
   }
 }
